pyproject.toml
[project]
name = "14-premio-sof-front"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Dornas",email = "gabrielbdornas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "streamlit (>=1.52.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]


[tool.poetry]
package-mode=false


[tool.taskipy.tasks]
server = { cmd = "streamlit run src/app.py", help = "Runs the server." }


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
